requirements.txt
Django~=5.1.7
djangorestframework~=3.15.2
dj-rest-auth~=7.0.1
nltk~=3.9.1